namespaced whole app
