+++ app/views/refinery/menus/admin/menu_links/_resource_link.html.erb
@@ -10,7 +10,7 @@
 	</div>
 
 	<div class="field">
-    <%= f.label :id_attribute, t('id_attribute', :scope => 'refinery.admin.menu_links') %>
+    <%= f.label :id_attribute, t('id_attribute', :scope => 'refinery.menus.admin.menu_links') %>
     <%= f.text_field :id_attribute %>
   </div>
