+++ app/views/refinery/menus/admin/menu_links/destroy.js.erb
