+++ app/views/refinery/menus/admin/menus/_form_actions.html.erb
@@ -1,23 +1,23 @@
 <div id='left_actions'>
   
   <div id="custom_url_box" class="pp-add-box">
-    <%= form_for [refinery, :admin, @menu.links.build], remote: true do |f| %>
+    <%= form_for [refinery, :menus, :admin, @menu.links.build], remote: true do |f| %>
       <%= f.hidden_field :refinery_menu_id %>
       <h3>Custom Link</h3>
       <div class='field'>
-        <%= f.label :custom_url, t('custom_url', scope: 'refinery.admin.menu_links.custom_link') %>
+        <%= f.label :custom_url, t('custom_url', scope: 'refinery.menus.admin.menu_links.custom_link') %>
         <%= f.text_field :custom_url %>
       </div>
       <div class='field'>
         <%= f.label :label %>
         <%= f.text_field :label %>
       </div>
-    <%= f.submit t('add', :scope => 'refinery.admin.menu_links.custom_link') %>
+    <%= f.submit t('add', :scope => 'refinery.menus.admin.menu_links.custom_link') %>
     <% end %>
   </div>
 
   <% Refinery::Menus.menu_resources.each do |resource_type, config| %>
-    <%= render partial: "refinery/admin/menu_links/resource_link_editor_box", locals: { resource_type: resource_type, config: config } %>
+    <%= render partial: "refinery/menus/admin/menu_links/resource_link_editor_box", locals: { resource_type: resource_type, config: config } %>
   <% end %>
 
 </div>
