+++ app/views/refinery/menus/admin/menus/_menu.html.erb
@@ -5,9 +5,9 @@
   </span>
   <span class='actions'>
     
-    <%= link_to refinery_icon_tag("application_edit.png"), refinery.edit_admin_menu_path(menu),
+    <%= link_to refinery_icon_tag("application_edit.png"), refinery.edit_menus_admin_menu_path(menu),
          :title => t('.edit') %>
-    <%= link_to refinery_icon_tag("delete.png"), refinery.admin_menu_path(menu),
+    <%= link_to refinery_icon_tag("delete.png"), refinery.menus_admin_menu_path(menu),
         :class => "cancel confirm-delete",
         :title => t('.delete'),
         :confirm => t('message', :scope => 'refinery.admin.delete', :title => menu.title),
