+++ app/views/refinery/menus/admin/menus/_records.html.erb
