+++ config/locales/en.yml
@@ -4,32 +4,33 @@ en:
       refinery_menus:
         title: Menus
         description: Create and manage custom menus for your site
-    admin:
-      menu_links:
-        remove: remove
-        add: "Add %{resource}"
-        custom_link:
-          add: Add Custom Link
-          custom_url: Url
-        id_attribute: ID attribute
-      menus:
-        form:
-          no_links: Use the left side to add links to your menu.
-        records:
-          title: Menus
-          no_items_yet: There are no menues yet. Ask your developer to create one for you.
-        menu:
-          edit: Edit this menu
-          delete: Remove this menu forever
+    menus:
+      admin:
+        menu_links:
+          remove: remove
+          add: "Add %{resource}"
+          custom_link:
+            add: Add Custom Link
+            custom_url: Url
+          id_attribute: ID attribute
+        menus:
+          form:
+            no_links: Use the left side to add links to your menu.
+          records:
+            title: Menus
+            no_items_yet: There are no menues yet. Ask your developer to create one for you.
+          menu:
+            edit: Edit this menu
+            delete: Remove this menu forever
   activerecord:
     models:
-      refinery/menu: Menu
-      refinery/menu_link: Link
+      refinery/menus/menu: Menu
+      refinery/menus/menu_link: Link
     attributes:
-      refinery/menu:
+      refinery/menus/menu:
         title: Title
         permatitle: Permatitle
-      refinery/menu_link:
+      refinery/menus/menu_link:
         title_attribute: Title attribute
         custom_url: Url
         label: Label
