+++ db/migrate/09_add_namespace_to_tables.rb
@@ -1,0 +1,6 @@
+class AddNamespaceToTables < ActiveRecord::Migration
+  def change
+    rename_table :refinery_menu_links, :refinery_menus_links
+  end
+
+end
