+++ lib/refinery/menu_item_decorator.rb
